+++ src/Controller/TeacherController.php
@@ -1,7 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Controller;
 
+use App\Entity\Plan;
 use App\Entity\Teacher;
 use App\Form\TeacherType;
 use App\Repository\TeacherRepository;
@@ -13,18 +16,28 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/teacher')]
 class TeacherController extends AbstractController
 {
-    #[Route('/', name: 'teacher_index', methods: ['GET'])]
-    public function index(TeacherRepository $teacherRepository): Response
+    #[Route('/plan/{plan}', name: 'teacher_index', methods: ['GET'])]
+    public function index(TeacherRepository $teacherRepository, Plan $plan): Response
     {
+        if ($this->getUser() != $plan->getUser()) {
+            return new Response('unauthorized', 401);
+        }
+
         return $this->render('teacher/index.html.twig', [
-            'teachers' => $teacherRepository->findAll(),
+            'teachers' => $teacherRepository->findBy(['plan' => $plan]),
+            'plan' => $plan
         ]);
     }
 
-    #[Route('/new', name: 'teacher_new', methods: ['GET', 'POST'])]
-    public function new(Request $request): Response
+    #[Route('/new/plan/{plan}', name: 'teacher_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, Plan $plan): Response
     {
+        if ($this->getUser() != $plan->getUser()) {
+            return new Response('unauthorized', 401);
+        }
+
         $teacher = new Teacher();
+        $teacher->setPlan($plan);
         $form = $this->createForm(TeacherType::class, $teacher);
         $form->handleRequest($request);
 
@@ -33,7 +46,7 @@ class TeacherController extends AbstractController
             $entityManager->persist($teacher);
             $entityManager->flush();
 
-            return $this->redirectToRoute('teacher_index');
+            return $this->redirectToRoute('teacher_index', ['plan' => $plan->getId()]);
         }
 
         return $this->render('teacher/new.html.twig', [
@@ -45,6 +58,10 @@ class TeacherController extends AbstractController
     #[Route('/{id}', name: 'teacher_show', methods: ['GET'])]
     public function show(Teacher $teacher): Response
     {
+        if ($this->getUser() != $teacher->getPlan()->getUser()) {
+            return new Response('unauthorized', 401);
+        }
+
         return $this->render('teacher/show.html.twig', [
             'teacher' => $teacher,
         ]);
@@ -53,13 +70,17 @@ class TeacherController extends AbstractController
     #[Route('/{id}/edit', name: 'teacher_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Teacher $teacher): Response
     {
+        if ($this->getUser() != $teacher->getPlan()->getUser()) {
+            return new Response('unauthorized', 401);
+        }
+
         $form = $this->createForm(TeacherType::class, $teacher);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('teacher_index');
+            return $this->redirectToRoute('teacher_index', ['plan' => $teacher->getPlan()->getId()]);
         }
 
         return $this->render('teacher/edit.html.twig', [
@@ -71,12 +92,18 @@ class TeacherController extends AbstractController
     #[Route('/{id}', name: 'teacher_delete', methods: ['POST'])]
     public function delete(Request $request, Teacher $teacher): Response
     {
+        if ($this->getUser() != $teacher->getPlan()->getUser()) {
+            return new Response('unauthorized', 401);
+        }
+
+        $plan = $teacher->getPlan();
+
         if ($this->isCsrfTokenValid('delete'.$teacher->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($teacher);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('teacher_index');
+        return $this->redirectToRoute('teacher_index', ['plan' => $plan->getId()]);
     }
 }
